+++ cmd/wallet_delete.go
@@ -19,12 +19,12 @@ var (
 	deleteWalletLong = cli.LongDesc(`
 		Delete the specified wallet and its keys.
 
-		Be sure to have a backup of its mnemonic, otherwise you won't be able to restore
-		it. If you lost it, you should transfer your funds, assets, orders, and anything else attached
-		to this wallet to another wallet.
+		Be sure to have its recovery phrase, otherwise you won't be able to restore it. If you
+		lost it, you should transfer your funds, assets, orders, and anything else attached to 
+		this wallet to another wallet.
 
-		The deletion removes the file in which the wallet and its keys are stored, meaning
-		you can reuse the wallet name, without causing any conflict.
+		The deletion removes the file in which the wallet and its keys are stored, meaning you
+		can reuse the wallet name, without causing any conflict.
 	`)
 
 	deleteWalletExample = cli.Examples(`