+++ wallet/hd_wallet_test.go
@@ -13,13 +13,13 @@ import (
 )
 
 const (
-	TestMnemonic1 = "swing ceiling chaos green put insane ripple desk match tip melt usual shrug turkey renew icon parade veteran lens govern path rough page render"
+	TestRecoveryPhrase1 = "swing ceiling chaos green put insane ripple desk match tip melt usual shrug turkey renew icon parade veteran lens govern path rough page render"
 )
 
 func TestHDWallet(t *testing.T) {
 	t.Run("Creating wallet succeeds", testHDWalletCreateWalletSucceeds)
 	t.Run("Importing wallet succeeds", testHDWalletImportingWalletSucceeds)
-	t.Run("Importing wallet with invalid mnemonic fails", testHDWalletImportingWalletWithInvalidMnemonicFails)
+	t.Run("Importing wallet with invalid recovery phrase fails", testHDWalletImportingWalletWithInvalidRecoveryPhraseFails)
 	t.Run("Importing wallet with unsupported version fails", testHDWalletImportingWalletWithUnsupportedVersionFails)
 	t.Run("Generating key pair succeeds", testHDWalletGeneratingKeyPairSucceeds)
 	t.Run("Generating key pair on isolated wallet fails", testHDWalletGeneratingKeyPairOnIsolatedWalletFails)
@@ -59,11 +59,11 @@ func testHDWalletCreateWalletSucceeds(t *testing.T) {
 	name := vgrand.RandomStr(5)
 
 	// when
-	w, mnemonic, err := wallet.NewHDWallet(name)
+	w, recoveryPhrase, err := wallet.NewHDWallet(name)
 
 	// then
 	require.NoError(t, err)
-	assert.NotEmpty(t, mnemonic)
+	assert.NotEmpty(t, recoveryPhrase)
 	assert.NotNil(t, w)
 	assert.Equal(t, uint32(2), w.Version())
 }
@@ -88,7 +88,7 @@ func testHDWalletImportingWalletSucceeds(t *testing.T) {
 			name := vgrand.RandomStr(5)
 
 			// when
-			w, err := wallet.ImportHDWallet(name, TestMnemonic1, tc.version)
+			w, err := wallet.ImportHDWallet(name, TestRecoveryPhrase1, tc.version)
 
 			// then
 			require.NoError(tt, err)
@@ -98,7 +98,7 @@ func testHDWalletImportingWalletSucceeds(t *testing.T) {
 	}
 }
 
-func testHDWalletImportingWalletWithInvalidMnemonicFails(t *testing.T) {
+func testHDWalletImportingWalletWithInvalidRecoveryPhraseFails(t *testing.T) {
 	tcs := []struct {
 		name    string
 		version uint32
@@ -121,7 +121,7 @@ func testHDWalletImportingWalletWithInvalidMnemonicFails(t *testing.T) {
 			w, err := wallet.ImportHDWallet(name, "vladimir harkonnen doesn't like trees", tc.version)
 
 			// then
-			require.ErrorIs(tt, err, wallet.ErrInvalidMnemonic)
+			require.ErrorIs(tt, err, wallet.ErrInvalidRecoveryPhrase)
 			assert.Nil(tt, w)
 		})
 	}
@@ -132,7 +132,7 @@ func testHDWalletImportingWalletWithUnsupportedVersionFails(t *testing.T) {
 	name := vgrand.RandomStr(5)
 
 	// when
-	w, err := wallet.ImportHDWallet(name, TestMnemonic1, 3)
+	w, err := wallet.ImportHDWallet(name, TestRecoveryPhrase1, 3)
 
 	// then
 	require.ErrorIs(t, err, wallet.NewUnsupportedWalletVersionError(3))
@@ -166,7 +166,7 @@ func testHDWalletGeneratingKeyPairSucceeds(t *testing.T) {
 			meta := []wallet.Meta{{Key: "env", Value: "test"}}
 
 			// when
-			w, err := wallet.ImportHDWallet(name, TestMnemonic1, tc.version)
+			w, err := wallet.ImportHDWallet(name, TestRecoveryPhrase1, tc.version)
 
 			// then
 			require.NoError(tt, err)
@@ -205,7 +205,7 @@ func testHDWalletGeneratingKeyPairOnIsolatedWalletFails(t *testing.T) {
 			name := vgrand.RandomStr(5)
 
 			// when
-			w, err := wallet.ImportHDWallet(name, TestMnemonic1, tc.version)
+			w, err := wallet.ImportHDWallet(name, TestRecoveryPhrase1, tc.version)
 
 			// then
 			require.NoError(tt, err)
@@ -255,7 +255,7 @@ func testHDWalletTaintingKeyPairSucceeds(t *testing.T) {
 			name := vgrand.RandomStr(5)
 
 			// when
-			w, err := wallet.ImportHDWallet(name, TestMnemonic1, tc.version)
+			w, err := wallet.ImportHDWallet(name, TestRecoveryPhrase1, tc.version)
 
 			// then
 			require.NoError(tt, err)
@@ -305,7 +305,7 @@ func testHDWalletTaintingKeyThatIsAlreadyTaintedFails(t *testing.T) {
 			name := vgrand.RandomStr(5)
 
 			// when
-			w, err := wallet.ImportHDWallet(name, TestMnemonic1, tc.version)
+			w, err := wallet.ImportHDWallet(name, TestRecoveryPhrase1, tc.version)
 
 			// then
 			require.NoError(tt, err)
@@ -361,7 +361,7 @@ func testHDWalletTaintingUnknownKeyFails(t *testing.T) {
 			name := vgrand.RandomStr(5)
 
 			// when
-			w, err := wallet.ImportHDWallet(name, TestMnemonic1, tc.version)
+			w, err := wallet.ImportHDWallet(name, TestRecoveryPhrase1, tc.version)
 
 			// then
 			require.NoError(tt, err)
@@ -396,7 +396,7 @@ func testHDWalletUntaintingKeyPairSucceeds(t *testing.T) {
 			name := vgrand.RandomStr(5)
 
 			// when
-			w, err := wallet.ImportHDWallet(name, TestMnemonic1, tc.version)
+			w, err := wallet.ImportHDWallet(name, TestRecoveryPhrase1, tc.version)
 
 			// then
 			require.NoError(tt, err)
@@ -460,7 +460,7 @@ func testHDWalletUntaintingKeyThatIsNotTaintedFails(t *testing.T) {
 			name := vgrand.RandomStr(5)
 
 			// when
-			w, err := wallet.ImportHDWallet(name, TestMnemonic1, tc.version)
+			w, err := wallet.ImportHDWallet(name, TestRecoveryPhrase1, tc.version)
 
 			// then
 			require.NoError(tt, err)
@@ -510,7 +510,7 @@ func testHDWalletUntaintingUnknownKeyFails(t *testing.T) {
 			name := vgrand.RandomStr(5)
 
 			// when
-			w, err := wallet.ImportHDWallet(name, TestMnemonic1, tc.version)
+			w, err := wallet.ImportHDWallet(name, TestRecoveryPhrase1, tc.version)
 
 			// then
 			require.NoError(tt, err)
@@ -546,7 +546,7 @@ func testHDWalletUpdatingKeyPairMetaSucceeds(t *testing.T) {
 			meta := []wallet.Meta{{Key: "primary", Value: "yes"}}
 
 			// when
-			w, err := wallet.ImportHDWallet(name, TestMnemonic1, tc.version)
+			w, err := wallet.ImportHDWallet(name, TestRecoveryPhrase1, tc.version)
 
 			// then
 			require.NoError(tt, err)
@@ -597,7 +597,7 @@ func testHDWalletUpdatingKeyPairMetaWithUnknownPublicKeyFails(t *testing.T) {
 			meta := []wallet.Meta{{Key: "primary", Value: "yes"}}
 
 			// when
-			w, err := wallet.ImportHDWallet(name, TestMnemonic1, tc.version)
+			w, err := wallet.ImportHDWallet(name, TestRecoveryPhrase1, tc.version)
 
 			// then
 			require.NoError(tt, err)
@@ -632,7 +632,7 @@ func testHDWalletDescribingPublicKeysSucceeds(t *testing.T) {
 			name := vgrand.RandomStr(5)
 
 			// when
-			w, err := wallet.ImportHDWallet(name, TestMnemonic1, tc.version)
+			w, err := wallet.ImportHDWallet(name, TestRecoveryPhrase1, tc.version)
 
 			// then
 			require.NoError(tt, err)
@@ -679,7 +679,7 @@ func testHDWalletDescribingUnknownPublicKeysFails(t *testing.T) {
 			name := vgrand.RandomStr(5)
 
 			// when
-			w, err := wallet.ImportHDWallet(name, TestMnemonic1, tc.version)
+			w, err := wallet.ImportHDWallet(name, TestRecoveryPhrase1, tc.version)
 
 			// then
 			require.NoError(tt, err)
@@ -715,7 +715,7 @@ func testHDWalletListingPublicKeysSucceeds(t *testing.T) {
 			name := vgrand.RandomStr(5)
 
 			// when
-			w, err := wallet.ImportHDWallet(name, TestMnemonic1, tc.version)
+			w, err := wallet.ImportHDWallet(name, TestRecoveryPhrase1, tc.version)
 
 			// then
 			require.NoError(tt, err)
@@ -766,7 +766,7 @@ func testHDWalletListingKeyPairsSucceeds(t *testing.T) {
 			name := vgrand.RandomStr(5)
 
 			// when
-			w, err := wallet.ImportHDWallet(name, TestMnemonic1, tc.version)
+			w, err := wallet.ImportHDWallet(name, TestRecoveryPhrase1, tc.version)
 
 			// then
 			require.NoError(tt, err)
@@ -819,7 +819,7 @@ func testHDWalletSigningTxSucceeds(t *testing.T) {
 			data := []byte("Je ne connaîtrai pas la peur car la peur tue l'esprit.")
 
 			// when
-			w, err := wallet.ImportHDWallet(name, TestMnemonic1, tc.version)
+			w, err := wallet.ImportHDWallet(name, TestRecoveryPhrase1, tc.version)
 
 			// then
 			require.NoError(tt, err)
@@ -865,7 +865,7 @@ func testHDWalletSigningTxWithTaintedKeyFails(t *testing.T) {
 			data := []byte("Je ne connaîtrai pas la peur car la peur tue l'esprit.")
 
 			// when
-			w, err := wallet.ImportHDWallet(name, TestMnemonic1, tc.version)
+			w, err := wallet.ImportHDWallet(name, TestRecoveryPhrase1, tc.version)
 
 			// then
 			require.NoError(tt, err)
@@ -915,7 +915,7 @@ func testHDWalletSigningTxWithUnknownKeyFails(t *testing.T) {
 			data := []byte("Je ne connaîtrai pas la peur car la peur tue l'esprit.")
 
 			// when
-			w, err := wallet.ImportHDWallet(name, TestMnemonic1, tc.version)
+			w, err := wallet.ImportHDWallet(name, TestRecoveryPhrase1, tc.version)
 
 			// then
 			require.NoError(tt, err)
@@ -962,7 +962,7 @@ func testHDWalletSigningAnyMessageSucceeds(t *testing.T) {
 			data := []byte("Je ne connaîtrai pas la peur car la peur tue l'esprit.")
 
 			// when
-			w, err := wallet.ImportHDWallet(name, TestMnemonic1, tc.version)
+			w, err := wallet.ImportHDWallet(name, TestRecoveryPhrase1, tc.version)
 
 			// then
 			require.NoError(tt, err)
@@ -1006,7 +1006,7 @@ func testHDWalletSigningAnyMessageWithTaintedKeyFails(t *testing.T) {
 			data := []byte("Je ne connaîtrai pas la peur car la peur tue l'esprit.")
 
 			// when
-			w, err := wallet.ImportHDWallet(name, TestMnemonic1, tc.version)
+			w, err := wallet.ImportHDWallet(name, TestRecoveryPhrase1, tc.version)
 
 			// then
 			require.NoError(tt, err)
@@ -1056,7 +1056,7 @@ func testHDWalletSigningAnyMessageWithUnknownKeyFails(t *testing.T) {
 			data := []byte("Je ne connaîtrai pas la peur car la peur tue l'esprit.")
 
 			// when
-			w, err := wallet.ImportHDWallet(name, TestMnemonic1, tc.version)
+			w, err := wallet.ImportHDWallet(name, TestRecoveryPhrase1, tc.version)
 
 			// then
 			require.NoError(tt, err)
@@ -1096,7 +1096,7 @@ func testHDWalletVerifyingAnyMessageSucceeds(t *testing.T) {
 			data := []byte("Je ne connaîtrai pas la peur car la peur tue l'esprit.")
 
 			// when
-			w, err := wallet.ImportHDWallet(name, TestMnemonic1, tc.version)
+			w, err := wallet.ImportHDWallet(name, TestRecoveryPhrase1, tc.version)
 
 			// then
 			require.NoError(tt, err)
@@ -1141,7 +1141,7 @@ func testHDWalletVerifyingAnyMessageWithUnknownKeyFails(t *testing.T) {
 			sig := []byte{0xd5, 0xc4, 0x9e, 0xfd, 0x13, 0x73, 0x9b, 0xdd, 0x36, 0x81, 0x75, 0xcc, 0x59, 0xc8, 0xbe, 0xe1, 0x20, 0x25, 0xe4, 0xb9, 0x14, 0x7a, 0x22, 0xbb, 0xa4, 0x84, 0xef, 0x7e, 0xe7, 0x2f, 0x55, 0x13, 0x5f, 0x52, 0x55, 0xad, 0x90, 0x35, 0x67, 0x6c, 0x91, 0x9d, 0xbb, 0x91, 0x21, 0x1f, 0x98, 0x53, 0xcc, 0x68, 0xe, 0x58, 0x5b, 0x4c, 0x26, 0xd7, 0xea, 0x20, 0x1, 0x50, 0x6c, 0x41, 0xcb, 0x3}
 
 			// when
-			w, err := wallet.ImportHDWallet(name, TestMnemonic1, tc.version)
+			w, err := wallet.ImportHDWallet(name, TestRecoveryPhrase1, tc.version)
 
 			// then
 			require.NoError(tt, err)
@@ -1180,7 +1180,7 @@ func testHDWalletMarshalingWalletSucceeds(t *testing.T) {
 			name := vgrand.RandomStr(5)
 
 			// when
-			w, err := wallet.ImportHDWallet(name, TestMnemonic1, tc.version)
+			w, err := wallet.ImportHDWallet(name, TestRecoveryPhrase1, tc.version)
 
 			// then
 			require.NoError(tt, err)
@@ -1226,7 +1226,7 @@ func testHDWalletMarshalingIsolatedWalletSucceeds(t *testing.T) {
 			name := vgrand.RandomStr(5)
 
 			// when
-			w, err := wallet.ImportHDWallet(name, TestMnemonic1, tc.version)
+			w, err := wallet.ImportHDWallet(name, TestRecoveryPhrase1, tc.version)
 
 			// then
 			require.NoError(tt, err)
@@ -1324,7 +1324,7 @@ func testHDWalletGettingWalletInfoSucceeds(t *testing.T) {
 			name := vgrand.RandomStr(5)
 
 			// when
-			w, err := wallet.ImportHDWallet(name, TestMnemonic1, tc.version)
+			w, err := wallet.ImportHDWallet(name, TestRecoveryPhrase1, tc.version)
 
 			// then
 			require.NoError(tt, err)
@@ -1356,7 +1356,7 @@ func testHDWalletGettingIsolatedWalletInfoSucceeds(t *testing.T) {
 			name := vgrand.RandomStr(5)
 
 			// when
-			w, err := wallet.ImportHDWallet(name, TestMnemonic1, tc.version)
+			w, err := wallet.ImportHDWallet(name, TestRecoveryPhrase1, tc.version)
 
 			// then
 			require.NoError(tt, err)
@@ -1403,7 +1403,7 @@ func testHDWalletIsolatingWalletSucceeds(t *testing.T) {
 	for _, tc := range tcs {
 		t.Run(tc.name, func(tt *testing.T) {
 			// when
-			w, err := wallet.ImportHDWallet(walletName, TestMnemonic1, tc.version)
+			w, err := wallet.ImportHDWallet(walletName, TestRecoveryPhrase1, tc.version)
 
 			// then
 			require.NoError(tt, err)
@@ -1447,7 +1447,7 @@ func testHDWalletIsolatingWalletWithTaintedKeyPairFails(t *testing.T) {
 			name := vgrand.RandomStr(5)
 
 			// when
-			w, err := wallet.ImportHDWallet(name, TestMnemonic1, tc.version)
+			w, err := wallet.ImportHDWallet(name, TestRecoveryPhrase1, tc.version)
 
 			// then
 			require.NoError(tt, err)
@@ -1496,7 +1496,7 @@ func testHDWalletIsolatingWalletWithNonExistingKeyPairFails(t *testing.T) {
 			name := vgrand.RandomStr(5)
 
 			// when
-			w, err := wallet.ImportHDWallet(name, TestMnemonic1, tc.version)
+			w, err := wallet.ImportHDWallet(name, TestRecoveryPhrase1, tc.version)
 
 			// then
 			require.NoError(tt, err)
@@ -1532,7 +1532,7 @@ func testHDWalletGettingWalletMasterKeySucceeds(t *testing.T) {
 			name := vgrand.RandomStr(5)
 
 			// when
-			w, err := wallet.ImportHDWallet(name, TestMnemonic1, tc.version)
+			w, err := wallet.ImportHDWallet(name, TestRecoveryPhrase1, tc.version)
 			require.NoError(tt, err)
 			require.NotNil(tt, w)